chore: rename components/web-view/collaborator-utils => utilities/plugin-utils
