+++ v3/src/components/case-table-card-common/case-table-card-title-bar.tsx
@@ -6,12 +6,12 @@ import TableIcon from "../../assets/icons/icon-table.svg"
 import { useDocumentContent } from "../../hooks/use-document-content"
 import { updateDataContextNotification } from "../../models/data/data-set-notifications"
 import { getTileDataSet } from "../../models/shared/shared-data-utils"
+import { getPreventDataContextReorg } from "../../utilities/plugin-utils"
 import { t } from "../../utilities/translation/translate"
 import { kCaseCardTileType } from "../case-card/case-card-defs"
 import { kCaseTableTileType } from "../case-table/case-table-defs"
 import { ComponentTitleBar } from "../component-title-bar"
 import { ITileTitleBarProps } from "../tiles/tile-base-props"
-import { getPreventDataContextReorg } from "../web-view/collaborator-utils"
 import { toggleCardTable } from "./case-table-card-utils"
 
 import "./case-table-card-title-bar.scss"
