+++ v3/src/components/case-table/collection-table.tsx
@@ -24,8 +24,8 @@ import { IDataSet } from "../../models/data/data-set"
 import { createAttributesNotification } from "../../models/data/data-set-notifications"
 import { uniqueName } from "../../utilities/js-utils"
 import { mstReaction } from "../../utilities/mst-reaction"
+import { getPreventParentCollectionReorg } from "../../utilities/plugin-utils"
 import { t } from "../../utilities/translation/translate"
-import { getPreventParentCollectionReorg } from "../web-view/collaborator-utils"
 import { useCaseTableModel } from "./use-case-table-model"
 import { useCollectionTableModel } from "./use-collection-table-model"
